Move TxIn to primitives
The `TxIn` has all public fields; move the `TxIn` to `primitives`.
