+++ primitives/src/lib.rs
@@ -59,6 +59,7 @@ pub use self::{
 pub use self::{
     locktime::{absolute, relative},
     witness::Witness,
+    transaction::TxIn,
 };
 
 #[rustfmt::skip]
